requirements.txt
numpy
Pillow
Requests
